@@ -14,7 +14,7 @@ description: Exemplo de integração Flutter com Lio - Cielo
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=3.0.0 < 4.0.0"
 
 dependencies:
   flutter:
@@ -22,9 +22,11 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  url_launcher: ^5.2.7
-  uni_links: ^0.2.0
+  cupertino_icons: ^1.0.8
+  url_launcher: ^6.3.1
+  app_links: ^6.4.0
+  go_router: ^14.8.1
+
 
 dev_dependencies:
   flutter_test:
