@@ -1,50 +1,50 @@
 class PaymentFields {
-  String cityState;
-  String serviceTax;
-  String documentType;
-  String hasConnectivity;
-  String document;
-  String entranceMode;
-  String paymentTypeCode;
-  String hasPassword;
-  String changeAmount;
-  String productName;
-  String hasWarranty;
-  String merchantName;
-  String hasSentMerchantCode;
-  String isOnlyIntegrationCancelable;
-  String primaryProductCode;
-  String originalTransactionId;
-  String requestDate;
-  String firstQuotaAmount;
-  String secondaryProductCode;
-  String pan;
-  String applicationName;
-  String isFinancialProduct;
-  String v40Code;
-  String merchantCode;
-  String totalizerCode;
-  String isDoubleFontPrintAllowed;
-  String hasSentReference;
-  String isExternalCall;
-  String originalTransactionDate;
-  String interestAmount;
-  String receiptPrintPermission;
-  String cardCaptureType;
-  String upFrontAmount;
-  String hasPrintedClientReceipt;
-  String creditAdminTax;
-  String primaryProductName;
-  String avaiableBalance;
-  String merchantAddress;
-  String applicationId;
-  String numberOfQuotas;
-  String boardingTax;
-  String paymentTransactionId;
-  String firstQuotaDate;
-  String hasSignature;
-  String secondaryProductName;
-  String statusCode;
+  String? cityState;
+  String? serviceTax;
+  String? documentType;
+  String? hasConnectivity;
+  String? document;
+  String? entranceMode;
+  String? paymentTypeCode;
+  String? hasPassword;
+  String? changeAmount;
+  String? productName;
+  String? hasWarranty;
+  String? merchantName;
+  String? hasSentMerchantCode;
+  String? isOnlyIntegrationCancelable;
+  String? primaryProductCode;
+  String? originalTransactionId;
+  String? requestDate;
+  String? firstQuotaAmount;
+  String? secondaryProductCode;
+  String? pan;
+  String? applicationName;
+  String? isFinancialProduct;
+  String? v40Code;
+  String? merchantCode;
+  String? totalizerCode;
+  String? isDoubleFontPrintAllowed;
+  String? hasSentReference;
+  String? isExternalCall;
+  String? originalTransactionDate;
+  String? interestAmount;
+  String? receiptPrintPermission;
+  String? cardCaptureType;
+  String? upFrontAmount;
+  String? hasPrintedClientReceipt;
+  String? creditAdminTax;
+  String? primaryProductName;
+  String? avaiableBalance;
+  String? merchantAddress;
+  String? applicationId;
+  String? numberOfQuotas;
+  String? boardingTax;
+  String? paymentTransactionId;
+  String? firstQuotaDate;
+  String? hasSignature;
+  String? secondaryProductName;
+  String? statusCode;
 
   PaymentFields({
     this.cityState,
